A lisp re-implementation of build-index.pl
One advantage is that this runs on my system in about 0.4 seconds
rather than the 10 seconds the Perl version takes. This difference is
actually due to the Perl version re-reading the files loads of times
to calculate byte offsets from line counts. I decided I could rewrite
the code in Lisp more easily than to alter the Perl (and I *think* I
was right).

Hopefully another benefit is that we can eventually move this
operation from build-time to run-time, which removes generation of
maxima-index.lisp entirely and means we needn't worry about whether it
is in sync with the info files.

The Perl code is very careful to calculate byte offsets rather than
character offsets. I'm much more cavalier (on purpose), since the
output of file-position (which I'm using) can be used as an input to
file-position to get to the same place. One worry is that if we build
with multiple lisps, they might disagree about what at file-position a
certain character is found. This would be a pain, but I don't want to
think hard to try and work around this, since hopefully this will all
run at run-time soon (hence with the correct lisp!)

Performance: On SBCL (with unicode), about 40% of the runtime is spent
in FILE-POSITION, since this works by essentially re-reading the file
up to the current point and doing the translation as we go. There
might be a performance improvement from using FILE-STRING-LENGTH and
keeping track ourselves, but I'm not convinced that will deal with
CR/CR+LF problems correctly (since READ-LINE doesn't include the
newline character)