Load offsets for info files at runtime rather than build time.
This moves info/build-index.lisp to src/parse-info.lisp and changes it
to key things by pathname rather than filename. Info files are parsed
the first time someone asks for help.

We touch a ridiculous number of files with this, but it's mostly just
removing references to maxima-info.lisp (thank goodness for git
grep...)