GCL doesn't have WILD-PATHNAME-P or PRINT-UNREADABLE-OBJECT
I'm not bothering with the wild-pathname-p check on GCL (hopefully
it'd never be triggered anyway).

This was found when testing correctness and timing on the various host
implementations. I ended up with the following table:

| impl  | load time /ms | total time /ms |
|-------+---------------+----------------|
| clisp |           124 |            319 |
| ecl   |           205 |            487 |
| gcl   |           114 |            197 |
| acl   |           247 |            579 |
| ccl   |           137 |            247 |
| cmucl |           155 |            390 |
| scl   |           137 |            360 |
| sbcl  |           101 |            247 |
(On a 2.3GHz x86 laptop)

Load time test:

   (in-package :cl-info)
   (progn
     (setf *documents* nil)
     (pushnew (info-pathname nil) *deferred-documents*)
     (time (topic-match "Number Theory" t)))

This tests how long we take to load up the top-level info file and
index. It's roughly the wait between the first time someone types "??
foo" and when they get the list of matching topics.

Total time test:

   (in-package :cl-info)
   (progn
     (setf *documents* nil)
     (pushnew (info-pathname nil) *deferred-documents*)
     (time (search-documentation-exact "factorial")))

This tests how long it takes to do the above, but also to load up a
different info file (I selected the topic to make sure of that) and
output it. If the user started his or her session with "? factorial",
this is how long he or she would have to wait to read the
documentation.

A note on the timings: ECL is always horrendously slow, so that's not
a huge surprise. I'm more surprised by the slow performance on ACL,
but I reckon it's acceptable for the following reasons. (1) Someone
running ACL will not be on an embedded device. (2) I don't feel the
need to optimise Maxima to run as fast as possible on a non-free
implementation.