Deal sensibly with locales and external formats.
This has a bit of nasty portability code (see *external-formats* in
init-cl.lisp and with-open-info-file in parse-info.lisp for the full
horror), which I've tested reasonably carefully with the lisps I have
on hand.

If a lisp implementation doesn't support a given external format or
doesn't support external formats at all (because it's GCL and not
really a common lisp implementation, gaargh) then we use a default
encoding. There shouldn't be errors, but the data might be garbage.

What isn't checked: I don't have an ECL with unicode support, so I
haven't checked that what I wrote works. It's what the manual says
though!

TODO: Warnings in configure.in if you select obviously bad
      combinations of documentation and implementation.