Make a unified way to call the various lisp implementations
These are defined in common.mk and called run_<foo>, for foo in the
set {clisp,cmucl,scl,acl,sbcl,gcl,ccl,ecl}. Then use it with

  $(call run_foo,'(+ 1 2)')

or the like. The lisp form isn't quoted in common.mk, so you want
either '' or "" around it to stop the shell getting confused.