Massively refactor the parse-info code to run at build time instead of run-time.
Basically, all the complicated texinfo parsing stuff is now in
lisp-utils/parse-info.lisp. This gets run at build time to create
offset files, which have names of the form

  maxima-info-offsets-<foo>.lisp

where <foo> is the name of the lisp implementation.

In these offset files, for each topic, we store the list

  (TITLE FILENAME START LEN)

where TITLE is the name of the node or index entry and START, LEN are
suitable for FILE-POSITION on the host lisp.

At run time, we load up these files (almost instantaneous - they're
pretty simple!) and use them to load up info contents.