Move locale-subdir to init-cl.lisp
This seems like a more sensible place for it. The only reason I didn't
put it there in the first place was an (abortive) attempt at running
the parse-info code in a fresh lisp image rather than within Maxima
and re-using some lisp files. Turned out to be a terrible idea...