Fix incorrect node / section look ups.
I had made several mistakes understanding how Texinfo worked and
they *almost* cancelled each other out...

I was mistakenly using the section title (stripped) for the node
name. This mostly turns out to work, but fails for one node in the
pt_BR documentation. Unfortunately, I misunderstood what was going on
and assumed that Texinfo was taking a prefix of the node title.

The new logic is simpler (and correct!) and just looks to see whether
a node contains a section and, if so, uses that one.