Modify DIMENSION-ARRAY to honor pre/post and sub/superscript properties. This appears to mostly work as intended, although it has not been carefully tested, and in particular it has not been tested with indices which are big enough to cause line breaks.
See tests/displa-prepost-subsuper-scripts-input and ...-output for
examples.

A known bug is that stuff like foo[bar]^baz is handled by DIMENSION-SUPERSCRIPT
which does its own subscript/superscript placement, and as a result,
pre/post and sub/superscript properties are not honored. Probably the
right thing to do is to handle exponents via DIMENSION-ARRAY (perhaps
renaming it to reflect its more general task of placing things).