In code to handle compressed info, add a convenience function read_compressed_file to read a compressed file and return its content as a string. Also adjust ASDF and defsystem config files as necessary for this.
Note that read_compressed_file is not completely general because it does not take
different encodings into account. It works OK for content which is 7-bit ASCII
since that is a subset of UTF-8, but I suspect it will not work for Latin-1
or other 8-bit, 1-byte encodings, although, for the record, I did not
actually try it.

Note also that even with read_compressed_file, we are not anywhere near able
to read documentation items from compressed info, since the index has byte offsets
plus numbers of characters to read. That works OK for file input, but as it
stands, we can jump to the appropriate place in a byte array, or extract the
appropriate number of characters from a string, but we cannot do both.