Replace defmfun with defun for non-user functions
Find all uses of defmfun for functions that are not user-exposed
functions (those not starting with a $), and use defun instead.

Basically done automatically with

for f in `find src -name '*.lisp'`
do
  sed -i .bak 's;(defmfun \([^$]\);(defun \1;' $f
done

But had to define a new macro named "defun-maclisp" to support some
non-user functions that use the maclisp nargs syntax.  These functions
are mformat-translate-open, mformat, map1, mmapcar, and outermap1.

Testsuite passes.