Import texindex version 6.5
We need to have our own version of texindex because the behavior has
changed.  In texindex 5.2, \entry macros with four args (topic and
subtopic)  would produce a a nice index with subtopics listed.

However, in later versions, this is no longer supported when the C
program was replaced with a gawk program.

Thus, we'll use our own copy and implement it ourselves.  Maybe
someday this will be folded back into the upstream texindex.