Use separate indices and names for functions and variables
To be able to make category links for functions and variables with the
same name, use separate indices for functions and variables.  Update
the macros to do this and also to prefix the category anchors with fn
or vr depending on whether the name was defined with deffn or defvr.