Handle duplicate deffnx names
When the deffnx function name matches the deffn function name, don't
produce the anchors and index entries because they would be
duplicated.

This is done by making m4_deffn define a macro that returns the
function name and having m4_deffnx check the deffn name with the
current deffnx name.  If they're the same, skip the anchor and index
stuff.