Update texinfo to 6.6dev (May 18, 2019)
Update texinfo to the 6.6dev version that supports @subentry for
creating multi-level indices that we use for the category indices.

Maybe this can also be used for the category indices in the info
version and html version?  That would probably change the html
version, but it might be consistent with the pdf version, then.