A backwards-compatible maxima.html
The documentation used to begin with the file maxima.html.
Now this file is named maxima_0.html. As there were objections
against this change perhaps it is best to add a file with the
old name that forwards the users to the new one.

(cherry picked from commit e0d454a8fa3527d35f5c9b1a49880c01ec9c0283)