Make gamma_incomplete_lower respect gamma_expand
Previously, gamma_incomplete_lower(a,z) would always expand if a were
an integer or half-integer.  Make gamma_incomplete_lower only do these
expansions when gamma_expand is true.

However, for compatibility with other uses of gamma-incomplete-lower,
we added an optional arg to indicate whether to expand or not.  If the
optional arg is not given, we expand, for backward compatibility of
previous callers that might have expected the expansions to happen.
If the arg is given, $gamma_expand is set to the given arg value.