Remove translate property for the obsolete DOLIST Maxima special form
There was an old commented-out Maxima special form named DOLIST (no
dollar sign) in Maxima 5.0 and before.  It was very different than
Common Lisp's DOLIST: it was essentially just an alias for MPROGN.

Although the DEFMSPEC definition was commented-out (and then removed
in commit 66a62cf7 in 2007), the translate property definition has
still been around.

So if had

  foo () := ?dolist (1)

then calling foo would cause a lisp error when interpreted but return
1 when translated.

Let's remove this.