Fix bogus translation of atan2 with one float and one rational argument
When a call to atan2 had one float and one rational argument, float
contagion should be applied (just like in the interpreted case) and
the special translation for float arguments should be used.

The problem is that the arguments in this special translation were not
being properly converted to floats before being passed to ATAN, so a
translated (Maxima) rational would cause a lisp error.

tests/rtest_translator.mac runs as expected, with new tests.