Bug fix: operatorp(expr,op) was inconsistent with operatorp(expr,[op])
If the second argument to operatorp was a list, then the comparison
was made with ALIKE1 via member; however, if the second argument was
not a list, then the comparison was made with EQUAL instead.

This has always been the case since operatorp was introduced in
commit 9a949ce5 in 2002.

No problems with the test suite or share test suite.