Add some error checking and docs
Add some simple error checking for the number of equations and
variables and for the coefficients.

The return code is also changed.  If ifgl1 is negative set return code
to iflg1. If iflg2 is not all 1 (normal return), set code to 1.
Otherwise, set code to 0, to indicate everything was fine.