Print docstring if exact match fails for documentation
When trying to describe something (via "? foo") and there's no exact
match, try to look up the docstring and return that.  We print a note
that the exact match failed but there's a docstring.  This is to let
the user know there is not such topic in the manual.  But perhaps the
symbol has some lisp documentation attached, so it's useful if we can
display that.

You can test this by doing "load(hompack)", and then "?
hompack_fixpdf".  There's no manual entry for hompack_fixpdf, but
there is a docstring for this function.