Change convergence to use polynomial value
Use the algorithms from cpoly to compute an error bound of the value
of a polynomial at a given point.  This error bound is used to
determine if the algorithm has converged to a solution.

To do this we added the functiosn:

* POLYEV to evaluate the polynomial at a point and returning the value
  of the polynomial and the partial sums of the horner method.
* ERREV-POLY computes the error bound in the value of the polynomial

ABERTH-CONVERGES-P is modified to use the error bound to determine
convergence.