Modify convergence criterion
Allow the convergence criterion to be either that the change in root
locations are small enough or that the value of the polynomial is less
than the bound on the value.

This is controlled by *aberth-use-offsets-for-convergence*.

We allow the max number of iterations to be changed vi
*aberth-max-iterations*.

Finally, if convergence fails and we've reached the max number of
iterations, a message is printed.