Scale the polynomial roots
Computes a scaling of the roots to reduce the dynamic range among the
roots.  This modifies the coefficients of the polynomial used to find
the roots.  The scaling is then applied to the result to get the
correct roots.

Really helps with product(x-10^-k,k,1,8).  Before, the algorithm
converged immediately with incorrect roots.  After, the roots are
accurate to 15 digits except 10^-7 which is accurate to about 10
digits.