Need reversealias property for grind and plot labels
Grind and plot needs the reversealias property to be able to print out
the right thing.  Without this, the result is the verb (%) form.  We
really want the noun (%) form.  This happens in msize-atom which will
strip the dollar sign off the noun form to produce the desired result.

Also added some comments on what the properties are need for.