Print correct exponent if bfloat rounds up
As reported on the maxima list
(https://sourceforge.net/p/maxima/mailman/message/37353663/), some bfloats
aren't printed with the correct exponent if the number needs to be
rounded up and the rounding causes the exponent to change.

The solution here is to look at the number of digits that would be
printed.  If it's greater than effective-printprec, then we must have
had to round the number up to the next power of 10, but we didn't
account for that in the exponent.  So, when the number of digits is
not the expected length, increase the exponent by the number of excess
digits.  We also need to remove the excess digits from the number so
that the printed result has the right number of digits (if bftrunc is
false).

A couple of simple tests are added to rtest3.mac.