Handle very small values for acsch.
For acsch(x) with x < least-positive-normalized-double-float, we would
get an unnecessary overflow.  We can fix this by using the logarc
formula for acsch and noting that for such small x, the result is
log(2) - log(x).  Use this to compute acsch for very small x.

Note that we then sinh(acsch(x)) signals an overflow for such small
x.  I think this is ok since for such small x, 1/x also overflows.