Experiment to use reader macro to embed maxima expressions in lisp
We implement a reader macro (from Stavros) to allow embedding Maxima
expressions in Lisp code to make it easier to write.

So we can write

  (with-maxima-expr ""sin(?phi) + 2*round(?phi/%pi)*elliptic_ec(?m)")

instead of

(add (ftake '%sin phi)
     (mul 2
	  (mul (ftake '%round (div phi '$%pi))
	       (ftake '%elliptic_ec m))))

We add this macro to various places in ellipt.lisp for testing.

However some tests fail and this needs to be fixed before we can
generally allow this feature.