Add macros for producing A&S and DLMF links
The macros urlaands and urldlmf produce nice URL links that we can
click on in the HTML and pdf manual.  The info manual has the links
shown so the user can just cut-n-paste them to a browser.

Modified a couple references to A&S (Special.texi) and DLMF
(nset.texi) to use the macros to see how it looks.