Update makeinfo version test to work with MacOS
The main problem was that on MacOS, expr didn't like the regexp
operator "+" in, say, "[0=9]+".  This is easily fixed by using the
equivalent "[0-9][0-9]*".

We also added some more checks to make sure that makeinfo_version is
not empty and that makeinfo_major and makeinfo_minor are not empty
either.  If any of these are empty, produce an error message.

Tested this on MacOS and verified that the warning messages are
produced (by using a faulty regexp for expr).