If makeinfo fails, exit build_html.sh
If makeinfo fails to build in build_html.sh, exit the script with an
error code from makeinfo so that make will stop building the docs.
This will make it obvious that something needs to be fixed so that we
can build the html docs.

Previously, the build didn't stop and it was hard to notice something
bad happened.

Also, we were inserting a section for the documentation categories,
assuming the previous node was named "Function and Variable Index".
Of course, this is incorrect for most other languages.  This will be
done in the <lang>/maxima.texi file with the correct language-specific
node name.