Undefine category macros for Russian manual.
Currently the Russian manual includes parts of the English manual
which uses the category macros.  But categories aren't currently working for
the Russian manual (many missing @refs).  To work around this, replace
the category macros with macros that do nothing.

We also need build_html.sh to add the --force option to makeinfo so
that the docs will be made despite the errors.  This is better than
nothing until we get this all sorted out.