Do not run lisp init files when building categories
When building the categories, the invocation of lisp should disallow
loading of the user's init files so that it doesn't interfere or
change how the categories are generated.

Currently only added support for cmucl, ecl, and ccl64; need to look
up the docs for the other lisps to find out how to disable loading of
user init files.