Support makeinfo versions with 3 parts and simplify computation
Test versions of makeinfo have version numbers like 6.8.90.  Update
configure to handle this so that we get a version of 6.8.90 instead of
8.90, which would be wrong.

We also simplify the tests for the version number by creating a
numerical value by multiply the major version by 100 and adding the
minor version.  This assumes that the minor version is never larger
than 99.  Then we can just compare integers.  Thus 5.0 is version 500,
and that's easy to compare against 5.1 (version 501).