Need to define $exptdispflag in globals.lisp
For some reason, if we don't do this, we get errors in the test suite
that `exptdispflag` is undefined.  I couldn't find it being defined
anywhere and don't see how the default value is set, but just add a
defmvar for it with a default value of T, as mentioned in the user
manual for `exptdispflag`.

Add a simple docstring for it too.

Now the test suite runs without problems.