Add evflag property using defmvar :properties.
Examine mlisp.lisp for where the evflag property is set and find the
corresponding variable in globals.lisp and use :properties to set the
evflag property.

While we're at it, add some docstrings for the variables that didn't
have any.  We basically just use a sentence or two from the user
manual.