Add some examples for how general polynomials are displayed
The previous version didn't show how maxima would display things like
assoc_legendre_p(n,m,x) as
```
      m
     P (x)
      n
```
So add an example for this and the other functions in
orthopoly.texi.m4.

Also added a note that you can't run `update_examples` on this file
because some examples depend on the previous example, which doesn't
work with `update_examples`.