First cut at having sinint.lisp not use specvar pardenom
It appears as if `aprog` is the main function that computes
`pardenom`.  Instead of having it produce this by side-effect, make
`aprog` return the computed value of `pardenom` as the second return
value.  This means we can rename `pardenom` in `aprog` to
`sinint-pardenom` so that we're sure we're not using the specvar.

Update all callers of `aprog` to look at the second returned value to
set `pardenom` appropriately.

(Note, there are some other changes to pass `pardenom` to other
functions, but I'm not sure they're needed.)

Full testsuite passes.