Add new file, global-var-init, for initializing globals.
Instead of initializing global vars in `initialize-runtime-globals`,
add a new file, "global-var-init.lisp" to initialize these.  Some
variables need special handling because we can't initialize them in
globals.lisp because the necessary functions aren't available when
globals.lisp is compiled.  Hence, have globals-var-init do the
initialization.  Move the initializations from
`initialize-runtime-globals` to global-var-init.lisp.

This request came from a private communication for a user (unixlisp)
who wanted to be able to load maxima without running `cl-user::run`.
This is not a use-case I care about, but moving the initialization to
a separate file works for this use-case as well as the normal usage of
maxima with a repl.